# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BRAMON API",
        "contact": {
            "name": "Support Team",
            "email": "support@bramonmeteor.org"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/public/captures": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "List all captures",
                "operationId": "/v1/public/captures",
                "responses": {
                    "200": {
                        "description": "List all captures",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/v1/public/captures/{capture}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "View a capture",
                "operationId": "/v1/public/captures/000-000-0000",
                "parameters": [
                    {
                        "name": "capture",
                        "in": "path",
                        "description": "The capture identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the capture details.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When capture not exists or not public."
                    }
                }
            }
        },
...
        "/v1/public/stations": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "List all stations",
                "operationId": "/v1/public/stations",
                "responses": {
                    "200": {
                        "description": "List all stations",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/v1/public/stations/{station}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "View a station",
                "operationId": "/v1/public/stations/000-000-0000",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "The station identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the station details.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When station not exists or not public."
                    }
                }
            }
        },
...
        "/v1/operator/operators": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "List all operators",
                "operationId": "/v1/operator/operators",
                "responses": {
                    "200": {
                        "description": "List all operators",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
        "/v1/operator/operators/{operator}": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "View an operator",
                "operationId": "/v1/operator/operators/000-000-0000",
                "parameters": [
                    {
                        "name": "operator",
                        "in": "path",
                        "description": "The operator identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the operator details.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When station not exists or not public."
                    }
                }
            },
            "put": {
                "tags": [
                    "Operators"
                ],
                "summary": "Update an operator",
                "operationId": "/v1/operator/operators/000-0000-00000",
                "parameters": [
                    {
                        "name": "operator",
                        "in": "path",
                        "description": "The operator identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The operator name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "The operator email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "city",
                        "in": "path",
                        "description": "The operator city",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "The operator state",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "mobile_phone",
                        "in": "path",
                        "description": "The operator mobile phone",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the capture data.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found. When capture not exists or not public."
                    },
                    "422": {
                        "description": "Error: Bad Param. When file not found or can not be deleted."
                    }
                }
            }
        },
...
    }
}